% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{annotatePeak,MeRIP.Peak-method}
\alias{annotatePeak,MeRIP.Peak-method}
\title{annotatePeak}
\usage{
\S4method{annotatePeak}{MeRIP.Peak}(object, threads = 1)
}
\arguments{
\item{MeRIP.Peak}{The MeRIP.Peak object}
}
\description{
annotatePeak
}
