% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{BetaBinTest}
\alias{BetaBinTest}
\title{BetaBinTest}
\usage{
\S4method{BetaBinTest}{MeRIP.Peak}(object, AdjIPeffi = TRUE,
  AdjustGC = FALSE, BSgenome = BSgenome.Hsapiens.UCSC.hg38)
}
\arguments{
\item{object}{The MeRIP.Peak object}

\item{AdjIPeffi}{Logic option determining whether to adjust overall IP efficiency, default is TRUE.}

\item{AdjustGC}{Logic option determining whether to adjust GC bias in each bins, default is FALSE.}

\item{BSgenome}{The BSgenome object that will used for calculating GC content for each bin if AdjustGC = TRUE.}
}
\description{
Perform inferential test using beta-binomial regression model.
}
