% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{callPeakBinomial}
\alias{callPeakBinomial}
\title{callPeakBinomial}
\usage{
callPeakBinomial(MeRIP, min_counts = 15, peak_cutoff_fdr = 0.05,
  peak_cutoff_oddRatio = 1, threads = 1)
}
\arguments{
\item{MeRIP}{The MeRIP object.}

\item{min_counts}{The minimal number of reads (input + IP) required in a bin to be called a peak.}

\item{peak_cutoff_fdr}{The cutoff of fdr to call a bin peak.}

\item{peak_cutoff_oddRatio}{The minimal oddRatio (IP/input) threshold to call a peak.}

\item{threads}{The number of threads to use. Default uses 1 threads.}
}
\description{
callPeakBinomial
}
