% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{callPeakFisher}
\alias{callPeakFisher}
\title{callPeakFisher}
\usage{
callPeakFisher(MeRIP, min_counts = 15, peak_cutoff_fdr = 0.05,
  peak_cutoff_oddRatio = 1, threads = 1)
}
\arguments{
\item{MeRIP}{The MeRIP object from countReads function.}

\item{min_counts}{The minimal number of reads present in a bin to be called a peak.}

\item{peak_cutoff_fdr}{The cutoff of fdr of fisher's exact test to call peak.}

\item{peak_cutoff_oddRatio}{The minimal oddRatio of fisher's exact test to call peak.}

\item{threads}{The number of threads to use. Default uses 1 threads.}
}
