% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{consistentPeak,MeRIP.Peak-method}
\alias{consistentPeak,MeRIP.Peak-method}
\title{consistentPeak}
\usage{
\S4method{consistentPeak}{MeRIP.Peak}(object, samplenames = NULL,
  joint_threshold = NA, threads = 1)
}
\arguments{
\item{object}{The MeRIP.Peak object contain peak calling result.}

\item{samplenames}{The samplenames to be reported for consistent peaks.}

\item{joint_threshold}{Define the number of sample required to have consistent peak in a locus to call consistent peak in a group.}
}
\value{
Peak consistent across specified samples at specified joint_threshod. If joint_threshold not specified, report consistent peaks across all samples specified. If no sample specified, report consistent peak across all samples.
}
