% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{counts,MeRIP-method}
\alias{counts,MeRIP-method}
\alias{counts,MeRIP.Peak-method}
\title{counts}
\usage{
\S4method{counts}{MeRIP}(object)

\S4method{counts}{MeRIP.Peak}(object)
}
\arguments{
\item{object}{The MeRIP object}
}
\description{
Extract all (Input + IP) read count of the MeRIP object.
}
