% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.Bino_test}
\alias{.Bino_test}
\title{Binomial test}
\usage{
.Bino_test(IP, input, IP_overall, input_overall, pseudo_count = 1)
}
\value{
data frame with p-values and odds ratio
}
\description{
Binomial test
}
