% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{geneBins,MeRIP-method}
\alias{geneBins,MeRIP-method}
\title{geneBins}
\usage{
\S4method{geneBins}{MeRIP}(object)
}
\arguments{
\item{object}{The MeRIP object}
}
\value{
data.frame
}
\description{
Extract the data.frame that maps the consecutive bins to genes.
}
