% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{geneExpression,MeRIP.Peak-method}
\alias{geneExpression,MeRIP.Peak-method}
\title{geneExpression}
\usage{
\S4method{geneExpression}{MeRIP.Peak}(object)
}
\arguments{
\item{object}{The MeRIP object}
}
\description{
Extract gene level expression (RNAseq) data in normalized read counts
}
