% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_logOR.R
\name{get_peak_logOR}
\alias{get_peak_logOR}
\title{get_peak_logOR}
\usage{
get_peak_logOR(MeRIPdata, vcf_file = NULL,
  BSgenome = BSgenome.Hsapiens.UCSC.hg19, AdjustGC = TRUE,
  AdjIPeffi = TRUE, thread = 1)
}
\arguments{
\item{MeRIPdata}{The MeRIP.Peak object containing peak calling result.}

\item{vcf_file}{The vcf file for genotype. The chromosome position must be sorted!!}

\item{BSgenome}{The BSgenome object. This needs to match the genome version of the gtf files.}

\item{AdjustGC}{Logic. Choose whether explicitly adjust GC bias (default: TRUE).}

\item{AdjIPeffi}{Logic. Choose whether explicitly adjust overall IP efficiency (default: TRUE).}
}
\description{
Compute peak logOR, adjust GC bias and IP efficiency for m6A QTL analysis
}
