% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuitarPlot2.R
\name{GuitarPlotNew}
\alias{GuitarPlotNew}
\title{GuitarPlotNew}
\usage{
GuitarPlotNew(gfeatures, GuitarCoordsFromTxDb = NA, txdb = NA,
  genome = NA, noBins = 10, saveToPDFprefix = NA, returnCount = FALSE,
  includeNeighborDNA = FALSE, maximalFeatureAmbiguity = 5,
  rescaleComponent = TRUE, fill = FALSE, adjust = 1)
}
