% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{Input.files}
\alias{Input.files}
\alias{Input.files,MeRIP-method}
\title{Input.files}
\usage{
Input.files(object)

\S4method{Input.files}{MeRIP}(object)
}
\arguments{
\item{object}{The MeRIP object}
}
\description{
Extract path to INPUT BAM files.
}
