% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{IP.files}
\alias{IP.files}
\alias{IP.files,MeRIP-method}
\title{IP.files}
\usage{
IP.files(object)

\S4method{IP.files}{MeRIP}(object)
}
\arguments{
\item{object}{The MeRIP object}
}
\description{
Extract path to IP BAM files.
}
