% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{jointPeakCount}
\alias{jointPeakCount}
\title{jointPeakCount}
\usage{
jointPeakCount(MeRIPdata)
}
\arguments{
\item{MeRIPdata}{The data list as output of callPeakFisher()}

\item{joint_threshold}{Require a bin to be called as a peak in joint_threshold number of samples to be a joint peak.}
}
\value{
Returns the MeRIPdata with joint peak read count for input and IP stored.
}
\description{
Extract read count of joint peaks for each sample and store the read counts in the MeRIP.Peak object.
}
