% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{normalizeLibrary,MeRIP.Peak-method}
\alias{normalizeLibrary,MeRIP.Peak-method}
\title{normalizeLibrary}
\usage{
\S4method{normalizeLibrary}{MeRIP.Peak}(object)
}
\arguments{
\item{object}{MeRIP.Peak object.}
}
\description{
Normalized the input as RNA-seq data and normalize IP by enrichment. Specifically, we normalize ip libraries sizes so that the geometry mean of enrichment are the same.
}
