% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{peakDistribution}
\alias{peakDistribution}
\title{plot distribution of peaks on gene annotation}
\usage{
\S4method{peakDistribution}{MeRIP.Peak}(object)
}
\arguments{
\item{object}{The MeRIP.Peak object}

\item{saveName}{the file name to save ditribution plot}
}
\description{
plot distribution of peaks on gene annotation
}
