% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetaGene.R
\name{plotMetaGeneMulti}
\alias{plotMetaGeneMulti}
\title{plotMEtaGeneMulti A wrapper function for Guitar to plot meta gene plot of multiple samples overlaid on each other.}
\usage{
plotMetaGeneMulti(peakList, gtf, saveToPDFprefix = NA,
  includeNeighborDNA = FALSE)
}
\arguments{
\item{peakList}{The list of peak, each object of list should have a data frame of peak in bed12 format.}

\item{gtf}{The annotation file for the gene model.}

\item{saveToPDFprefix}{Set a name to save the plot to PDF.}

\item{includeNeighborDNA}{Whether to include upstrean and downstream region in the meta gene.}
}
