% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.DESeq2.R
\name{plotPCAfromMatrix}
\alias{plotPCAfromMatrix}
\title{plotPCAfromMatrix}
\usage{
plotPCAfromMatrix(m, group, standardize = TRUE, loglink = TRUE)
}
\arguments{
\item{m}{The matrix of count data}

\item{group}{The factor levels to color the samples. Should be the save number as the # of matrix columns}

\item{standardize}{Logic parameter indicating whether to standardize the count data to have unit variance. The default is TRUE.}

\item{loglink}{Logic parameter determine whether to take log of the metrix data. Default is TRUE. If your input matrix is at log scale, use FALSE.}
}
\description{
plotPCAfromMatrix
}
