% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{PrepCoveragePlot,MeRIP.Peak-method}
\alias{PrepCoveragePlot,MeRIP.Peak-method}
\title{Prepare coverage plot}
\usage{
\S4method{PrepCoveragePlot}{MeRIP.Peak}(object, gtf = NULL)
}
\arguments{
\item{object}{The MeRIP object}

\item{gtf}{optional gtf file if the stored path to gtf file has changed.}
}
\description{
import GTF into the MeRIP object for plot
}
