% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{QNBtest}
\alias{QNBtest}
\title{Wraper function to use QNB package to test for differential peaks.}
\usage{
\S4method{QNBtest}{MeRIP.Peak}(object)
}
\arguments{
\item{object}{The MeRIP.Peak object}
}
