% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.pvalue.R
\name{qqplot.pvalue}
\alias{qqplot.pvalue}
\title{qqplot.pvalue}
\usage{
qqplot.pvalue(x, pointSize = 1, legendSize = 4)
}
\arguments{
\item{x}{can be a vector (p value of one group) or a list of vector (p value of multiple groups).}

\item{pointSize}{The size of data points}

\item{legendSize}{The size of points in the legend}
}
