% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_betaBinomial2.R
\name{QTL_BetaBin2}
\alias{QTL_BetaBin2}
\title{QTL_BetaBin2}
\usage{
QTL_BetaBin2(MeRIPdata, vcf_file, BSgenome = BSgenome.Hsapiens.UCSC.hg19,
  testWindow = 1e+05, Chromosome, Range = NULL, Covariates = NULL,
  AdjustGC = TRUE, PCsToInclude = 0, normalizeGenotype = FALSE,
  thread = 1)
}
\arguments{
\item{MeRIPdata}{The MeRIP.Peak object}

\item{vcf_file}{The vcf file for genotype. The chromosome position must be sorted!!}

\item{BSgenome}{The BSgenome object. This needs to match the genome version of the gtf files.}

\item{testWindow}{Integer. Test SNPs in <testWindow> bp window flanking the peak.}

\item{Chromosome}{The chromsome to run QTL test.}

\item{Range}{The position range on a chromosome to test.}

\item{Covariates}{The matrix for covariates to be included in the test.}

\item{AdjustGC}{Logic. Choose whether explicitly adjust GC bias.}

\item{normalizeGenotype}{Logic. Choose whether genotype is normalized to mean = 0, var = 1 before regression.}
}
\description{
Same as QTL_BetaBin except sizeFactor estimated by Median of ratio method was used to scale library size instead of total coverage.
}
