% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_PoissonGamma.R
\name{QTL_PoissonGamma}
\alias{QTL_PoissonGamma}
\title{QTL_PoissonGamma}
\usage{
QTL_PoissonGamma(pheno, vcf_file, peak_bed, testWindow = 1e+05, Chromosome,
  Range = NULL, Covariates = NULL, maxPsi = 100, thread = 1)
}
\arguments{
\item{pheno}{The phenotype data matrix. Needs to be IP read count that has been normalized for expression level.}

\item{vcf_file}{The vcf file for genotype. The chromosome position must be sorted!!}

\item{peak_bed}{The peak file in BED12 format that needs to correspond to phenotype data matrix.}

\item{testWindow}{Integer. Test SNPs in <testWindow> bp window flanking the peak.}

\item{Chromosome}{The chromsome to run QTL test.}

\item{Range}{The position range on a chromosome to test.}

\item{Covariates}{The matrix for covariates to be included in the test.}

\item{maxPsi}{The max estimation for the random effect parameter Psi.}
}
