% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_betaBinomial.neg.permute.R
\name{swapChr22.QTL_BetaBin.permute}
\alias{swapChr22.QTL_BetaBin.permute}
\title{swapChr22.QTL_BetaBin.permute}
\usage{
swapChr22.QTL_BetaBin.permute(MeRIPdata, vcf_file,
  BSgenome = BSgenome.Hsapiens.UCSC.hg19, testWindow = 1e+05, Chromosome,
  Range = NULL, Covariates = NULL, AdjustGC = TRUE, AdjIPeffi = TRUE,
  PCsToInclude = 0, thread = 1, Nround = 1)
}
\arguments{
\item{MeRIPdata}{The MeRIP.Peak object}

\item{vcf_file}{The vcf file for genotype. The chromosome position must be sorted!!}

\item{BSgenome}{The BSgenome object. This needs to match the genome version of the gtf files.}

\item{testWindow}{Integer. Test SNPs in <testWindow> bp window flanking the peak.}

\item{Chromosome}{The chromsome to run QTL test.}

\item{Range}{The position range on a chromosome to test.}

\item{Covariates}{The matrix for covariates to be included in the test.}
}
