% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{jointPeakCount}
\alias{jointPeakCount}
\title{jointPeakCount}
\usage{
jointPeakCount(MeRIPdata)
}
\arguments{
\item{MeRIPdata}{The data list as output of callPeakFisher()}
}
\value{
Returns the MeRIP object with joint peak read count for Input and IP stored.
}
\description{
Extract read count of joint peaks for each sample and store the read counts in the MeRIP.Peak object.
}
