% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{RADARtest}
\alias{RADARtest}
\title{Perform inferential test using Poisson Random effect model in RADAR}
\usage{
\S4method{RADARtest}{MeRIP.Peak}(object, exclude = NULL, maxPsi = 100)
}
\arguments{
\item{object}{The MeRIP.Peak object}
}
